Add BookDetails loader and rename loaders to match team convention
